Move Euclidean pattern helpers to libfp
